Resolve open code-scanning alerts
The sheathed-object save interpolated the owner name and the serialized
object text into its INSERT, so a quote in any saved object text, such
as a restrung description, broke the statement and the sheathed item was
not saved. Bind the row through a prepared statement as the main object
save does, keeping its refusal to store a truncated payload. The three
player object-save transactions repeated a delete by name, and the rent
header update also interpolated the name; both now escape it with
mysql_escape_string_alloc. The object-save database test now saves a
sheath row with hostile text under both SQL modes and checks that
oversized payloads are not stored.

say_spell spliced the metamagic flourish into the start-message format
string. Copy the flourish into the message ahead of the formatted text
instead. A test watching the message showed the flourish itself was
wrong: cast_spell recorded the cast's metamagic only after the start
message, so a first metamagic cast showed no flourish and a later plain
cast showed the previous one. Record it before the message. The new test
watches a metamagic cast and a following plain cast.

Widen the zcheck experience limit and the mission reward products to
long before multiplying, and test a remaining duration of more than one
round as duration > 0 rather than duration + 1 > 1.

+++ src/act.wizard.c
@@ -5754,7 +5754,7 @@ ACMD(do_links)
 #define MAX_DAMROLL_ALLOWED MAX(GET_LEVEL(mob) / 5, 1)
 #define MAX_HITROLL_ALLOWED MAX(GET_LEVEL(mob) / 3, 1)
 #define MAX_MOB_GOLD_ALLOWED GET_LEVEL(mob) * 20
-#define MAX_EXP_ALLOWED GET_LEVEL(mob) * GET_LEVEL(mob) * 120
+#define MAX_EXP_ALLOWED ((long)GET_LEVEL(mob) * GET_LEVEL(mob) * 120)
 #define MAX_LEVEL_ALLOWED LVL_IMPL
 #define GET_OBJ_AVG_DAM(obj) (((GET_OBJ_VAL(obj, 2) + 1) / 2.0) * GET_OBJ_VAL(obj, 1))
 /* arbitrary limit for per round dam */
@@ -5988,7 +5988,7 @@ ACMD(do_zcheck)
                               GET_GOLD(mob), MAX_MOB_GOLD_ALLOWED);
 
       if (GET_EXP(mob) > MAX_EXP_ALLOWED && (found = 1))
-        len = snprintf_append(buf, sizeof(buf), (int)len, "- Has %ld experience (limit: %d)\r\n",
+        len = snprintf_append(buf, sizeof(buf), (int)len, "- Has %ld experience (limit: %ld)\r\n",
                               GET_EXP(mob), MAX_EXP_ALLOWED);
       if ((AFF_FLAGGED(mob, AFF_CHARM) || AFF_FLAGGED(mob, AFF_POISON)) && (found = 1))
         len = snprintf_append(buf, sizeof(buf), (int)len,
